Moved music controller buttons to it's own class
